Fix SPI and GPIO handling to support recent kernels.
This driver was previously using the SYSFS API to interact with GPIO. This API is deprecated since kernel 4.8 but was still working up to 5.10. The SYSFS API was removed in more recent version and is now replaced by a character device interface.

In this commit, I replaced calls to the old SYSFS API to this new GPIO interface. I also added a small test program to test GPIO (display info about all GPIO chips and tries to detect the gpiochip corresponding to the CH341 of the USB adapter).

Also, the initialization of the SPI bus now binds to spidev driver since this is needed to make /dev/spidevx.y available since kernel 5.1.